refactor: move HTTP processing logic from plugin to client library
Move HTTP method field processing and body conversion functions
from actions/http/main.go to http/client.go for better testability
and separation of concerns.

Changes:
- Add ResolveMethodAndURL, ConvertBodyToFormEncoded, PrepareRequestData to http/client.go
- Update Request function to use PrepareRequestData
- Simplify HTTP action plugin to use client library functions directly
- Remove duplicate functions from plugin
