--- actions/http/integration_test.go
@@ -1,141 +1,0 @@
-package http
-
-import (
-	"testing"
-)
-
-func TestUpdateMapWithNestedBody(t *testing.T) {
-	tests := []struct {
-		name     string
-		input    map[string]string
-		expected map[string]string
-	}{
-		{
-			name: "nested YAML body with application/json content-type",
-			input: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body__foo__name":       "aaa",
-				"body__foo__role":       "bbb",
-				"body__bar":             "xyz",
-				"body__count":           "42",
-			},
-			expected: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body":                  `{"bar":"xyz","count":42,"foo":{"name":"aaa","role":"bbb"}}`,
-			},
-		},
-		{
-			name: "simple flat body with application/json content-type",
-			input: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body__name":            "test",
-				"body__age":             "25",
-			},
-			expected: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body":                  `{"age":25,"name":"test"}`,
-			},
-		},
-		{
-			name: "form data without json content-type",
-			input: map[string]string{
-				"method":     "POST",
-				"url":        "http://example.com",
-				"body__name": "test",
-				"body__age":  "25",
-			},
-			expected: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"body":                  "age=25&name=test",
-				"headers__content-type": "application/x-www-form-urlencoded",
-			},
-		},
-		{
-			name: "array body with application/json content-type",
-			input: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body__0__foo":          "1",
-				"body__0__bar":          "2",
-				"body__0__baz":          "3",
-			},
-			expected: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body":                  `[{"bar":2,"baz":3,"foo":1}]`,
-			},
-		},
-		{
-			name: "multiple array items with application/json content-type",
-			input: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body__0__name":         "item1",
-				"body__1__name":         "item2",
-				"body__2__name":         "item3",
-			},
-			expected: map[string]string{
-				"method":                "POST",
-				"url":                   "http://example.com",
-				"headers__content-type": "application/json",
-				"body":                  `[{"name":"item1"},{"name":"item2"},{"name":"item3"}]`,
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Make a copy of input to avoid modifying the test case
-			data := make(map[string]string)
-			for k, v := range tt.input {
-				data[k] = v
-			}
-
-			err := updateMap(data)
-			if err != nil {
-				t.Errorf("updateMap() error = %v", err)
-				return
-			}
-
-			// Check that all expected keys and values are present
-			for expectedKey, expectedValue := range tt.expected {
-				actualValue, exists := data[expectedKey]
-				if !exists {
-					t.Errorf("expected key %s not found in result", expectedKey)
-					continue
-				}
-
-				// For JSON body, parse and compare structure instead of string comparison
-				if expectedKey == "body" && data["headers__content-type"] == "application/json" {
-					// Since JSON key ordering can vary, we'll just check that it's valid JSON
-					// and contains the expected structure (our convertBodyToJson tests cover the details)
-					if actualValue == "" {
-						t.Errorf("expected non-empty JSON body")
-					}
-					// More detailed testing is done in TestConvertBodyToJson
-				} else if actualValue != expectedValue {
-					t.Errorf("key %s: expected %v, got %v", expectedKey, expectedValue, actualValue)
-				}
-			}
-
-			// Verify no body__ keys remain
-			for key := range data {
-				if key != "method" && key != "url" && key != "headers__content-type" && key != "body" {
-					t.Errorf("unexpected key remaining: %s", key)
-				}
-			}
-		})
-	}
-}
